+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 This changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.0] - 2024-07-29
+### Added
+- Support for Python 3.11
+
+### Fixed
+- All logs are now debug level
+- Endpoint addresses removed from logs
+
 # [0.6.0] - 2023-03-29
 ### Added
 - FallbackProvider.